refactor: restructure components and add scroll hooks
