+++ src/components/common/FullScrollSection.tsx
