+++ src/components/common/ScrollSectionTrigger.tsx
