+++ src/components/common/ScrollVideoSection.tsx
@@ -9,11 +9,8 @@ import {
   isValidElement,
   useState,
 } from 'react'
-import gsap from 'gsap'
-import { ScrollTrigger } from 'gsap/ScrollTrigger'
 import type { OverlayProps } from './Overlay'
-
-gsap.registerPlugin(ScrollTrigger)
+import { useScrollVideo } from '@/hooks/useScrollVideo'
 
 type Props = {
   src: string
@@ -35,67 +32,15 @@ export default function ScrollVideoSection({
   const wrapperRef = useRef<HTMLDivElement>(null)
   const [currentTime, setCurrentTime] = useState(0)
 
-  // Initialize scroll-triggered video control when activated
-  useEffect(() => {
-    if (!activateTriggers) return
-
-    const video = videoRef.current
-    const container = containerRef.current
-    const wrapper = wrapperRef.current
-    if (!video || !container || !wrapper) return
-
-    const killAll = () => {
-      ScrollTrigger.getAll().forEach(trigger => trigger.kill())
-      gsap.globalTimeline.clear()
-    }
-
-    let tween: gsap.core.Tween | null = null
-    const duration = video.duration
-
-    if (!isFinite(duration) || duration === 0) return
-    const height = duration * scrollSpeed
-
-    // Reset all triggers and animations
-    killAll()
-
-    // Scroll-linked video progress
-    tween = gsap.to(video, {
-      currentTime: duration,
-      ease: 'none',
-      scrollTrigger: {
-        trigger: container,
-        start: 'top top',
-        end: `+=${height}`,
-        scrub: true,
-        pin: true,
-        anticipatePin: 1,
-        onUpdate: () => setCurrentTime(video.currentTime),
-      },
-    })
-
-    // Optional visual zoom effect based on scroll position
-    ScrollTrigger.create({
-      trigger: container,
-      start: 'top top',
-      end: `+=${height}`,
-      scrub: true,
-      onUpdate: (self) => {
-        const p = self.progress
-        if (p < 0.15 || p > 0.85) {
-          gsap.to(wrapper, { scale: 0.8, overwrite: 'auto' })
-        } else {
-          gsap.to(wrapper, { scale: 0.95, opacity: 1, overwrite: 'auto' })
-        }
-      }
-    })
-
-    ScrollTrigger.refresh()
-
-    return () => {
-      killAll()
-      tween?.kill()
-    }
-  }, [activateTriggers, scrollSpeed, src])
+  // Initialize scroll-triggered video logic using custom hook
+  useScrollVideo({
+    videoRef,
+    wrapperRef,
+    containerRef,
+    activate: activateTriggers,
+    speed: scrollSpeed,
+    onTimeUpdate: setCurrentTime,
+  })
 
   // Notify parent once video metadata is ready
   useEffect(() => {
