+++ src/components/presentation/HeroSection.tsx
