+++ src/components/presentation/RecapSection.tsx
@@ -2,7 +2,7 @@
 
 import { motion } from 'framer-motion'
 import { Download } from 'lucide-react'
-import SectionHeader from '@/components/SectionHeader'
+import SectionHeader from '@/components/common/SectionHeader'
 
 export default function RecapSection() {
   return (
